made the configSettings test work again
